moved common components into respective subcategories
